A-CodeQuality Completed
Some codes do not follow the coding standard.

Refactor the packages in the asher package to follow the package naming convention.
Refactor the java files inside test.java.asher to follow the class naming convention.
Rephrase method/class header comments to follow the format specified in the coding standard.
Java Docs are being added to most methods and classes for clarity.
