+++ src/main/java/asher/commands/DeleteCommand.java
@@ -1,15 +1,32 @@
-package asher.Commands;
+package asher.commands;
 
-import asher.Ui.Ui;
-import asher.Tasks.TaskList;
+import asher.ui.Ui;
+import asher.tasks.TaskList;
 import asher.BotException;
-import asher.Tasks.Task;
+import asher.tasks.Task;
 
+/**
+ * Represents a command to delete a task.
+ */
 public class DeleteCommand extends Command {
+
+    /**
+     * Constructs a Delete Command object with the given input string.
+     *
+     * @param input The input string for the Delete Command.
+     */
     public DeleteCommand(String input) {
         super(input);
     }
 
+    /**
+     * Executes the Delete Command.
+     *
+     * @param ui The UI object to interact with the user.
+     * @param taskList The list of tasks.
+     * @param storage The storage object for reading and writing tasks from a file.
+     * @return The String input for Delete Command.
+     */
     @Override
     public String execute(Ui ui, TaskList taskList, Storage storage) {
         try {
