+++ src/main/java/asher/commands/UnmarkCommand.java
@@ -1,15 +1,31 @@
-package asher.Commands;
+package asher.commands;
 
 import asher.BotException;
-import asher.Tasks.TaskList;
-import asher.Ui.Ui;
+import asher.tasks.TaskList;
+import asher.ui.Ui;
 
+/**
+ * Represents a command to unmark a task.
+ */
 public class UnmarkCommand extends Command {
 
+    /**
+     * Constructs a Unmark Command object with the given input string.
+     *
+     * @param input The input string for the Unmark Command.
+     */
     public UnmarkCommand(String input) {
         super(input);
     }
 
+    /**
+     * Executes the Unmark Command.
+     *
+     * @param ui The UI object to interact with the user.
+     * @param taskList The list of tasks.
+     * @param storage The storage object for reading and writing tasks from a file.
+     * @return The string input for Unmark Command.
+     */
     @Override
     public String execute(Ui ui, TaskList taskList, Storage storage) {
         try {
