+++ src/test/java/asher/tasks/TaskListTest.java
@@ -1,10 +1,10 @@
-package asher.Tasks;
+package asher.tasks;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
-public class taskListTest {
+public class TaskListTest {
     @Test
     public void testAddAndGetTask() {
         TaskList taskList = new TaskList();
