+++ src/test/java/asher/tasks/TaskTest.java
@@ -1,11 +1,11 @@
-package asher.Tasks;
+package asher.tasks;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class taskTest {
+public class TaskTest {
     @Test
     public void testMarkDoneAndUndone() {
         Task task = new Task("Task");
